re-work home page structure
