+++ src/components/features/home/AboutSwitcher.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const AboutSwitcher = () => {
+  return <div>switcher...</div>;
+};
